Refactor wal_dir_filter_builder into async_file and rename.
